audio_processing/requirements.txt
torch
numpy
openai-whisper
pyaudio

